Corrected max reconnects value handling
git-svn-id: file:///home/svn/repos/tigase-server/trunk@354 7d282ba1-3ae6-0310-8f9b-c9008a0864d2

+++ src/main/java/tigase/server/ConnectionManager.java
@@ -462,9 +462,9 @@ public abstract class ConnectionManager extends AbstractMessageReceiver
 			} catch (ConnectException e) {
 				// Accept side for component service is not ready yet?
 				// Let's wait for a few secs and try again.
-				Long reconnects = (Long)port_props.get(MAX_RECONNECTS_PROP_KEY);
+				Integer reconnects = (Integer)port_props.get(MAX_RECONNECTS_PROP_KEY);
 				if (reconnects != null) {
-					long recon = reconnects.longValue();
+					int recon = reconnects.intValue();
 					if (recon != 0) {
 						port_props.put(MAX_RECONNECTS_PROP_KEY, (--recon));
 						reconnectService(port_props, connectionDelay);
